requirements.txt
# Core ML dependencies
tensorflow>=2.15.0,<2.17.0
numpy>=1.21.0,<2.0.0
matplotlib>=3.5.0
Pillow>=8.0.0
pandas>=1.3.0

# Optional dependencies for dataset management
fiftyone>=0.21.0
opencv-python>=4.5.0

# Development and visualization
seaborn>=0.11.0
scikit-learn>=1.0.0